Phase B9: split desired_state/mod.rs into topical sub-modules
Extracts load, validate, convert, diff, normalize, and tests into
separate files under desired_state/; mod.rs retains only type defs
and re-exports. All 46 tests pass unchanged.
